Move DefaultCommandLine
