+++ subprojects/core-exec/src/main/java/dev/nokee/core/exec/DefaultCommandLine.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.core.exec.internal;
+package dev.nokee.core.exec;
 
 import dev.nokee.core.exec.CommandLine;
 import dev.nokee.core.exec.CommandLineTool;
